+++ CHANGE_LOG.md
@@ -2,10 +2,11 @@
 > Breaking Change
 - Rename `Cli` namespace to `Terminal`
 - Rename `CliOptions` to `TerminalOptions`
+- Rename `CliBuilder` to `TerminalBuilder`
 
 # 4.4.2-rc*
 > Breaking Change
-- Rename all classes with Cli* to Terminal*
+- Rename all classes with `Cli*` to `Terminal*`
 - Rename Router Remote Options
 - Check maximum command string length in router